refactor: Topbar에서 유저 프로필 분리, 유저 프로필 디자인 리뷰 반영
