+++ features/Topbar/UserProfile/index.ts
@@ -1,0 +1,1 @@
+export { UserProfile } from "./UserProfile";
